move plans around
